// font_end_khachhang/src/types/product.type.ts
export interface Product {
  id: number;
  name: string;
}
// ...
export interface IDetailProduct {
  code: string;
  color: string;
  id: number;
  images: any;
  index: number;
  name: string;
  price: number;
  quantity: number;
  size: string | number;
  sole: string;
  status: boolean;
  weight: number;
  discountPercent: number;
  discountValue?: number;
}
export interface IInforShoe {
  createAt: string;
  updateAt: string;
  createBy: string;
  updateBy: string;
  deleted: boolean;
  id: number;
  name: string;
  thuongHieu: Product;
  xuatXu: Product;
   tayAo: Product;
    coAo: Product;
     chatLieu: Product;

  description: any;
  
}